tests/test_10_cb_debug_tracing.py
"""
Unit tests for 10_cb_debug_tracing.py
Tests logging and OpenTelemetry tracing functionality.
"""
import unittest
from unittest.mock import Mock, patch, MagicMock, mock_open
import logging
from datetime import timedelta
import sys
import os


class TestDebugTracing(unittest.TestCase):
    
    def setUp(self):
        """Set up test fixtures."""
        self.mock_cluster = Mock()
        self.mock_bucket = Mock()
        self.mock_collection = Mock()
        self.mock_logger = Mock()
        self.mock_tracer = Mock()
    
    @patch('logging.basicConfig')
    @patch('logging.getLogger')
    def test_logging_configuration(self, mock_get_logger, mock_basic_config):
        """Test logging configuration setup."""
        # Test that we can verify logging configuration calls
        mock_basic_config.assert_not_called()  # Initially not called
        mock_get_logger.assert_not_called()    # Initially not called
        
        # Test the expected configuration
        expected_config = {
            'filename': 'couchbase_example.log',
            'filemode': 'w',
            'level': logging.DEBUG,
            'format': '%(levelname)s::%(asctime)s::%(message)s',
            'datefmt': '%Y-%m-%d %H:%M:%S'
        }
        
        # Verify the configuration values are correct
        self.assertEqual(expected_config['filename'], 'couchbase_example.log')
        self.assertEqual(expected_config['level'], logging.DEBUG)
    
    def test_couchbase_logging_configuration(self):
        """Test Couchbase SDK logging configuration."""
        with patch('couchbase.configure_logging') as mock_configure:
            # Mock the configure_logging function
            mock_logger = Mock()
            mock_logger.name = 'test_logger'
            mock_logger.level = logging.DEBUG
            
            # Call the function
            import couchbase
            couchbase.configure_logging(mock_logger.name, level=mock_logger.level)
            
            # Verify it was called
            mock_configure.assert_called_once_with(mock_logger.name, level=mock_logger.level)
    
    def test_opentelemetry_tracer_setup(self):
        """Test OpenTelemetry tracer setup."""
        with patch('opentelemetry.trace.set_tracer_provider') as mock_set_provider, \
             patch('opentelemetry.trace.get_tracer') as mock_get_tracer, \
             patch('opentelemetry.sdk.trace.TracerProvider') as mock_tracer_provider:
            
            # Mock tracer provider
            mock_provider_instance = Mock()
            mock_tracer_provider.return_value = mock_provider_instance
            
            # Mock tracer
            mock_tracer_instance = Mock()
            mock_get_tracer.return_value = mock_tracer_instance
            
            # Test the setup
            from opentelemetry import trace
            from opentelemetry.sdk.trace import TracerProvider
            
            provider = TracerProvider()
            trace.set_tracer_provider(provider)
            tracer = trace.get_tracer(__name__)
            
            # Verify calls
            mock_set_provider.assert_called_once()
            mock_get_tracer.assert_called_once_with(__name__)
    
    def test_span_processor_setup(self):
        """Test span processor setup."""
        with patch('opentelemetry.trace.get_tracer_provider') as mock_get_provider, \
             patch('opentelemetry.sdk.trace.export.SimpleSpanProcessor') as mock_processor, \
             patch('opentelemetry.sdk.trace.export.ConsoleSpanExporter') as mock_exporter:
            
            # Mock components
            mock_provider_instance = Mock()
            mock_get_provider.return_value = mock_provider_instance
            mock_processor_instance = Mock()
            mock_processor.return_value = mock_processor_instance
            mock_exporter_instance = Mock()
            mock_exporter.return_value = mock_exporter_instance
            
            # Test setup
            from opentelemetry import trace
            from opentelemetry.sdk.trace.export import SimpleSpanProcessor, ConsoleSpanExporter
            
            provider = trace.get_tracer_provider()
            processor = SimpleSpanProcessor(ConsoleSpanExporter())
            provider.add_span_processor(processor)
            
            # Verify calls
            mock_get_provider.assert_called_once()
            mock_exporter.assert_called_once()
            mock_processor.assert_called_once_with(mock_exporter_instance)
    
    def test_cluster_connection_setup(self):
        """Test cluster connection setup."""
        with patch('couchbase.cluster.Cluster') as mock_cluster_class, \
             patch('couchbase.options.ClusterOptions') as mock_options, \
             patch('couchbase.auth.PasswordAuthenticator') as mock_auth:
            
            # Mock components
            mock_auth_instance = Mock()
            mock_auth.return_value = mock_auth_instance
            mock_options_instance = Mock()
            mock_options.return_value = mock_options_instance
            mock_cluster_instance = Mock()
            mock_cluster_class.return_value = mock_cluster_instance
            
            # Test connection setup
            from couchbase.auth import PasswordAuthenticator
            from couchbase.cluster import Cluster
            from couchbase.options import ClusterOptions
            
            auth = PasswordAuthenticator("Administrator", "password")
            options = ClusterOptions(auth)
            cluster = Cluster('couchbase://your-ip', options)
            
            # Verify calls
            mock_auth.assert_called_once_with("Administrator", "password")
            mock_options.assert_called_once_with(mock_auth_instance)
            mock_cluster_class.assert_called_once_with('couchbase://your-ip', mock_options_instance)
    
    def test_wait_until_ready_configuration(self):
        """Test wait_until_ready configuration."""
        from couchbase.options import WaitUntilReadyOptions
        from couchbase.diagnostics import ServiceType
        from datetime import timedelta
        
        # Test WaitUntilReadyOptions creation
        options = WaitUntilReadyOptions(service_types=[ServiceType.KeyValue, ServiceType.Query])
        
        # Verify it was created successfully
        self.assertIsInstance(options, WaitUntilReadyOptions)
    
    def test_service_type_configuration(self):
        """Test ServiceType configuration in WaitUntilReadyOptions."""
        from couchbase.diagnostics import ServiceType
        from couchbase.options import WaitUntilReadyOptions
        
        # Test service types
        expected_services = [ServiceType.KeyValue, ServiceType.Query]
        options = WaitUntilReadyOptions(service_types=expected_services)
        
        # Verify the options object was created
        self.assertIsInstance(options, WaitUntilReadyOptions)
    
    def test_get_operation_with_tracing(self):
        """Test get operation with tracing span."""
        with patch('opentelemetry.trace.get_tracer') as mock_get_tracer:
            # Mock tracer and span
            mock_span = Mock()
            mock_tracer = Mock()
            mock_tracer.start_as_current_span.return_value.__enter__ = Mock(return_value=mock_span)
            mock_tracer.start_as_current_span.return_value.__exit__ = Mock(return_value=None)
            mock_get_tracer.return_value = mock_tracer
            
            # Mock collection and exception
            mock_collection = Mock()
            mock_collection.get.side_effect = Exception("Document not found")
            
            # Test the operation
            from opentelemetry import trace
            tracer = trace.get_tracer(__name__)
            
            with tracer.start_as_current_span("get_operation"):
                try:
                    mock_collection.get('not-a-key')
                except Exception as e:
                    self.assertEqual(str(e), "Document not found")
            
            # Verify tracer was called
            mock_get_tracer.assert_called_once_with(__name__)
    
    def test_tracer_decorator_functionality(self):
        """Test tracer decorator functionality."""
        with patch('opentelemetry.trace.get_tracer') as mock_get_tracer:
            # Mock tracer
            mock_tracer = Mock()
            mock_get_tracer.return_value = mock_tracer
            
            # Create a mock decorator
            def mock_decorator(span_name):
                def decorator(func):
                    def wrapper(*args, **kwargs):
                        return func(*args, **kwargs)
                    return wrapper
                return decorator
            
            mock_tracer.start_as_current_span = mock_decorator
            
            # Test decorator usage
            @mock_tracer.start_as_current_span("couchbase_operations")
            def test_function():
                return "test_result"
            
            result = test_function()
            self.assertEqual(result, "test_result")
    
    def test_logging_info_message(self):
        """Test logging info message."""
        with patch('logging.getLogger') as mock_get_logger:
            # Mock logger
            mock_logger = Mock()
            mock_get_logger.return_value = mock_logger
            
            # Test logging
            import logging
            logger = logging.getLogger()
            logger.info('Cluster is ready for operations.')
            
            # Verify logger was called
            mock_logger.info.assert_called_once_with('Cluster is ready for operations.')
    
    def test_couchbase_exception_handling(self):
        """Test CouchbaseException handling in perform_couchbase_operations."""
        # Mock CouchbaseException
        class MockCouchbaseException(Exception):
            pass
        
        # Test exception handling
        try:
            raise MockCouchbaseException("Test exception")
        except MockCouchbaseException as e:
            self.assertEqual(str(e), "Test exception")
    
    def test_cluster_connection_in_perform_operations(self):
        """Test cluster connection within perform_couchbase_operations."""
        with patch('couchbase.cluster.Cluster') as mock_cluster_class:
            # Mock cluster
            mock_cluster_instance = Mock()
            mock_cluster_class.return_value = mock_cluster_instance
            
            # Test cluster creation
            from couchbase.cluster import Cluster
            from couchbase.options import ClusterOptions
            from couchbase.auth import PasswordAuthenticator
            
            cluster = Cluster('couchbase://your-ip', 
                            ClusterOptions(PasswordAuthenticator("Administrator", "password")))
            
            # Verify cluster was created
            mock_cluster_class.assert_called_once()
    
    def test_main_function_execution(self):
        """Test main function execution."""
        # Test that the main block would execute correctly
        with patch('builtins.__name__', '__main__'):
            # Mock the perform_couchbase_operations function
            def mock_perform_operations():
                return "operations_complete"
            
            result = mock_perform_operations()
            self.assertEqual(result, "operations_complete")


if __name__ == '__main__':
    unittest.main()
