requirements.txt
couchbase==4.4.0
pandas>=1.5.0
openpyxl>=3.0.0
opentelemetry-api>=1.15.0
opentelemetry-sdk>=1.15.0